Fix the handling of non-provided paths
Some wrong phpdoc was omitting the possibility for the source name to be
null when a path is not provided, which led to missing it in a few
places. With PHP 8.1 deprecating passing null to a bunch of core
functions, this was detected.

(cherry picked from commit 195bd263a865ac13727eca7c6cfdcda369ca7a77)